Refactor directory structure of Feathub Java Udf
